% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohensd.R
\name{cohensd}
\alias{cohensd}
\title{Cohen's D coversion from OR}
\usage{
cohensd(p0, or)
}
\arguments{
\item{p0}{Proportion of the unexposed who had event of interest}

\item{or}{Odds Ratio}
}
\value{

}
\description{
Cohen's D coversion from OR
}
\examples{
cohensd(.01,2.0)
}
