% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{DATASET}
\alias{DATASET}
\title{Data set needed for example to run}
\format{
A dataset containing event indicator and exposure group
\describe{
  \item{Wave}{Expsoure group, categorical}
  \item{new_date}{date of test, date}
  \item{Days_to_testing}{number of days to test, numerical}
  \item{Event}{Event indicator, binary}

}
}
\usage{
DATASET
}
\description{
Data set needed for example to run
}
\keyword{datasets}
