% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsratio.R
\name{oddsratio}
\alias{oddsratio}
\title{Odds Ratio Function}
\usage{
oddsratio(unexposed, exposed)
}
\arguments{
\item{unexposed}{The proportion of those who are unexposed and had event}

\item{exposed}{The proportion of those who are exposed and had event}
}
\value{
The output from \code{\link{print}}
}
\description{
Use the prevalence of exposed and unexposed to calculate the OR.
}
\examples{
oddsratio(.01,0.9)
}
