% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p0.R
\name{p0}
\alias{p0}
\title{Proportion of those who experienced events are are unexposed}
\usage{
p0(event, exposure, exposure2)
}
\arguments{
\item{event}{Event Column name}

\item{exposure}{Exposure column name}

\item{exposure2}{Criteria of interest}
}
\value{

}
\description{
Proportion of those who experienced events are are unexposed
}
\examples{
p0(event = DATASET$Event, exposure = DATASET$Wave, exposure2 = "Delayed Incentive")
}
